pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flextrain"
version = "0.1.0"
description = "Fault-tolerant distributed training framework with async checkpointing"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Varun Dudipala"}
]
keywords = ["deep-learning", "distributed-training", "pytorch", "fsdp", "fault-tolerance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "aiosqlite>=0.19.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
]
gcs = ["google-cloud-storage>=2.10.0"]
s3 = ["boto3>=1.28.0"]

[project.scripts]
flextrain = "flextrain.cli.main:app"

[project.urls]
Homepage = "https://github.com/Varun-Dudipala/flextrain"
Repository = "https://github.com/Varun-Dudipala/flextrain"

[tool.setuptools.packages.find]
where = ["."]
include = ["flextrain*"]
